Add files via upload
This is an PDF file reader project in short we can say it as audio book which dictate all the text written in PDF file.
It also provide a benefit of dictating from a specific page in a PDF file. It will extract all the "TEXT" from the PDF file and read it aloud.